Refs #1354 Extract dashboard app
