+++ src/pollypm/cockpit_apps/__init__.py
@@ -1,5 +1,6 @@
 """Focused Textual app modules extracted from ``cockpit_ui``."""
 
+from pollypm.cockpit_apps.dashboard import PollyDashboardApp
 from pollypm.cockpit_apps.pane import PollyCockpitPaneApp
 
-__all__ = ["PollyCockpitPaneApp"]
+__all__ = ["PollyCockpitPaneApp", "PollyDashboardApp"]
